cleaned up #include conflicts

+++ card.h
@@ -1,11 +1,12 @@
 #pragma once
-#include <string>
+#include "Transform.h"
+
 #include "cinder/app/MouseEvent.h"
 #include "cinder/app/App.h"
 #include "cinder/Log.h"
 //#include "cinder/Area.h"
-#include "Transform.h"
 
+#include <string>
 
 using namespace ci;
 using namespace ci::app;
@@ -18,18 +19,18 @@ class Card
 {
 public:
 	Card();
-	Card(int, int, const int, const int);
+	Card(float, float, const float, const float);
 	~Card();
 	void update();
 	//State getcurrentstate();
 	void mouseDrag(MouseEvent event);
 	void mouseDown(MouseEvent event);
 
-	int x;
-	int y;
+	float x;
+	float y;
 	Rectf rect;
-	int width;
-	int height;
+	float width;
+	float height;
 	string title;
 	string subtitle;
 	string body;
@@ -44,7 +45,7 @@ public:
 
 	Transform transform;
 
-	void setpos(int m, int n);
+	void setpos(float m, float n);
 
 	/*
 	string rubrik
